+++ springboot-app-products/src/main/java/com/eucaliptus/springboot_app_products/model/Unit.java
@@ -15,7 +15,7 @@ public class Unit {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_unit")
-    private Long idUnit;
+    private int idUnit;
 
     @Column(name = "unit_name", nullable = false)
     private String unitName;
